[BO] Fix wrong day of the week in task list
Since day zero (epoch time) was a Thurday (Thu, 01 Jan 1970 00:00:00)
when using mktime() we need to take that into consideration.
Thursday is day 4, Monday is day 1; to make our counting start from 1
we need to decrement by 3.

This explains why after selecting a day when creating or editing a task
the day shown in the list was wrong.